+++ src/core/SkillMatrix/service/SkillMatrixService.ts
@@ -80,9 +80,20 @@ export class SkillMatrixService {
   }
 
   async deleteSkillMatrixByCompany(companyDomain: string): Promise<void> {
-    const allUsersCompany = await this.userProfileService.getByCompany(companyDomain)
-    for (const user of allUsersCompany) {
-      await this.skillMatrixRepository.delete(user.uid)
+    const originalData = await this.skillMatrixRepository.getData()
+    if (originalData?.length) {
+      const allUsersCompany = await this.userProfileService.getByCompany(companyDomain)
+      for (const user of allUsersCompany) {
+        try {
+          await this.skillMatrixRepository.delete(user.uid)
+        } catch (error) {
+          console.error("Error: delete all users of a company", error)
+          for (const item of originalData) {
+            this.skillMatrixRepository.restoreData(item)
+          }
+          break;
+        }
+      }
     }
   }
